requirements_utf8.txt
# ???쒕쾭 (FastAPI)
fastapi
uvicorn
# ?뚯씪 ?낅줈??OCR?? 泥섎━
python-multipart
flask-cors

# AI 諛??대씪?곕뱶 ?쒕퉬??
openai
python-dotenv

# PDF 諛??곗씠??泥섎━
pdf2image
pillow

# AWS RDS(PostgreSQL) ?곌껐
psycopg2-binary
# ?ㅼ씠踰??대줈諛?API ?몄텧
requests

#firebase
firebase-admin

apscheduler

# npm install -g firebase-tools

supabase

# pip install supabase
# pip install postgrest==0.10.6

# fastapi
# pydantic
# supabase
# uvicorn

